+++ src/main/java/com/nuc/zmblog/resp/TypeResp.java
@@ -16,5 +16,5 @@ public class TypeResp {
 
     private String name;
 
-    private Integer blogSize;
+    private Long blogSize;
 }
